test(tools): consolidate test files to one per source file
