+++ internal/tools/markdown_test.go
@@ -1,0 +1,156 @@
+package tools
+
+import (
+	"strings"
+	"testing"
+
+	"github.com/jmrplens/libgen-mcp/internal/libgen"
+)
+
+// TestResultIdentifier covers the doi and empty arms of resultIdentifier that the
+// md5-keyed search fixtures never reach.
+func TestResultIdentifier(t *testing.T) {
+	if got := resultIdentifier(libgen.Result{MD5: "abc"}); got != "md5:abc" {
+		t.Errorf("md5 identifier = %q, want %q", got, "md5:abc")
+	}
+	if got := resultIdentifier(libgen.Result{DOI: "10.1/x"}); got != "doi:10.1/x" {
+		t.Errorf("doi identifier = %q, want %q", got, "doi:10.1/x")
+	}
+	if got := resultIdentifier(libgen.Result{}); got != "" {
+		t.Errorf("empty identifier = %q, want empty", got)
+	}
+}
+
+// TestResultLinks covers the skip-empty-URL and default-label arms of resultLinks.
+func TestResultLinks(t *testing.T) {
+	// An entry with no URL is skipped; an entry with no label renders as "download".
+	r := libgen.Result{Downloads: []libgen.DownloadOption{
+		{Label: "GET", URL: ""},            // skipped: empty URL
+		{Label: "", URL: "https://m/dl/2"}, // default label "download"
+	}}
+	if got := resultLinks(r); got != "[download](https://m/dl/2)" {
+		t.Errorf("resultLinks = %q, want %q", got, "[download](https://m/dl/2)")
+	}
+	// No downloads at all → empty string.
+	if got := resultLinks(libgen.Result{}); got != "" {
+		t.Errorf("resultLinks(no downloads) = %q, want empty", got)
+	}
+}
+
+// TestSearchLinksSurfacedAndHinted verifies the search markdown table renders
+// each result's download links as Markdown links, and that the structured
+// next_steps carries the instruction to include those links in the reply.
+func TestSearchLinksSurfacedAndHinted(t *testing.T) {
+	out := SearchOutput{
+		Mirror: "m", Page: 1,
+		Results: []libgen.Result{{
+			Title: "A Book", MD5: "0123456789abcdef0123456789abcdef",
+			Downloads: []libgen.DownloadOption{{Label: "GET", URL: "https://mirror/dl/1"}},
+		}},
+	}
+	out.NextSteps = searchNextSteps(out)
+
+	md := renderSearchMarkdown(out)
+	if !strings.Contains(md, "Download links") {
+		t.Errorf("table should have a Download links column; got:\n%s", md)
+	}
+	if !strings.Contains(md, "[GET](https://mirror/dl/1)") {
+		t.Errorf("table should render the download link; got:\n%s", md)
+	}
+	steps := strings.Join(out.NextSteps, "\n")
+	if !strings.Contains(steps, "download links") {
+		t.Errorf("next_steps should instruct the model to include download links; got %q", steps)
+	}
+
+	// No links → no preserve-links hint.
+	noLinks := SearchOutput{Mirror: "m", Page: 1, Results: []libgen.Result{{Title: "B", MD5: "abc"}}}
+	if resultsHaveLinks(noLinks.Results) {
+		t.Fatal("fixture should have no links")
+	}
+	if strings.Contains(strings.Join(searchNextSteps(noLinks), "\n"), "download links") {
+		t.Error("next_steps should not mention download links when results carry none")
+	}
+}
+
+// TestRenderMarkdownEdgeCases covers the empty-search, doi-only details, and
+// resumed-download rendering branches.
+func TestRenderMarkdownEdgeCases(t *testing.T) {
+	empty := renderSearchMarkdown(SearchOutput{Mirror: "m", NextSteps: []string{"broaden it"}})
+	if !strings.Contains(empty, "No results") || !strings.Contains(empty, "broaden it") {
+		t.Errorf("empty search markdown should note no results and next steps; got:\n%s", empty)
+	}
+
+	details := renderDetailsMarkdown(DetailsOutput{
+		Edition:   map[string]any{"title": "Paper", "doi": "10.1/z"},
+		NextSteps: []string{"download it"},
+	})
+	if !strings.Contains(details, "Paper") || !strings.Contains(details, "10.1/z") {
+		t.Errorf("details markdown should show title and doi; got:\n%s", details)
+	}
+
+	dl := renderDownloadMarkdown(DownloadOutput{
+		DownloadResult: libgen.DownloadResult{Path: "/p", SizeBytes: 9, Source: "libgen", Resumed: true},
+	})
+	if !strings.Contains(dl, "Resumed") {
+		t.Errorf("resumed download markdown should note the resume; got:\n%s", dl)
+	}
+}
+
+// TestRenderDetails_BibtexFenceIsBreakoutSafe proves a BibTeX value carrying a
+// code-fence sequence cannot close the block early. renderDetailsMarkdown must
+// open the fence with more backticks than the longest backtick run inside the
+// content (the CommonMark closing-fence rule), so the injected "```" and any
+// trailing Markdown/instructions stay inside the fenced code block.
+func TestRenderDetails_BibtexFenceIsBreakoutSafe(t *testing.T) {
+	const bib = "@book{x,\n  title = {evil ``` ## Fake instruction},\n}"
+	out := renderDetailsMarkdown(DetailsOutput{
+		File:      map[string]any{"title": "Paper", "md5": "abc"},
+		Citations: &Citations{BibTeX: bib},
+	})
+
+	// Locate the opening fence: the first line after the "Citation (BibTeX)"
+	// heading that is a run of backticks (optionally followed by the info string).
+	var fence string
+	for line := range strings.SplitSeq(out, "\n") {
+		if strings.HasPrefix(line, "```") {
+			fence = line
+			break
+		}
+	}
+	if fence == "" {
+		t.Fatalf("no opening fence found:\n%s", out)
+	}
+	openLen := len(fence) - len(strings.TrimLeft(fence, "`"))
+
+	// The longest backtick run inside the content is 3 ("```"); the opening fence
+	// must be strictly longer so the content can never close it.
+	if openLen <= 3 {
+		t.Errorf("opening fence (%d backticks) must exceed the interior run (3):\n%s", openLen, out)
+	}
+	// The forged instruction must remain inside the block, never on its own
+	// top-level line as rendered Markdown.
+	if strings.Contains(out, "\n## Fake instruction") {
+		t.Errorf("injected heading broke out of the fence:\n%s", out)
+	}
+}
+
+// TestWriteEnrichment_UserFacingLabels verifies the enrichment markdown uses
+// user-facing labels (Journal, Times cited, Published year) rather than Crossref
+// jargon, and renders OpenLibrary fields too.
+func TestWriteEnrichment_UserFacingLabels(t *testing.T) {
+	out := renderDetailsMarkdown(DetailsOutput{
+		File: map[string]any{"md5": "d48739b6ac9e01d70dda1de46805d797"},
+		Enrichment: &libgen.Enrichment{
+			Crossref:    &libgen.CrossrefWork{ContainerTitle: "Cell", PublishedYear: 2011, CitationCount: 56374},
+			OpenLibrary: &libgen.OLBook{OpenLibURL: "https://openlibrary.org/works/OL1W", Description: "A classic."},
+		},
+	})
+	for _, want := range []string{"Journal / container: Cell", "Times cited: 56374", "Published year: 2011", "OpenLibrary record:", "A classic."} {
+		if !strings.Contains(out, want) {
+			t.Errorf("enrichment markdown should contain %q; got:\n%s", want, out)
+		}
+	}
+	if strings.Contains(out, "Crossref container") {
+		t.Error("enrichment markdown should not use the old 'Crossref container' jargon")
+	}
+}
